move DEVELOPMENT_TEAM into untracked Signing.xcconfig
Since I have a different DEVELOPMENT_TEAM than zvonler, having the DEVELOPMENT_TEAM setting in the pbxproj is annoying.

This change removes the DEVELOPMENT_TEAM setting from the pbxproj and adds a reference to Signing.xcconfig to the pbxproj. The Signing.xcconfig file is ignored
so changes to it won't be committed.

You need to create the Signing.xcconfig yourself. For zvonler:

    echo DEVELOPMENT_TEAM = JFYM87G8N9 > Signing.xcconfig

For mayoff:

    echo DEVELOPMENT_TEAM = GVBKTM899S > Signing.xcconfig